labs/lab_1/requirements.txt
certifi==2020.12.5
numpy==1.20.1
pandas==1.2.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
